Handle APIVersions Request 17

+++ app/main.js
@@ -7,7 +7,7 @@ const server = net.createServer((conn) => {
     console.log(data);
 
     let api_version_bytes = data.subarray(6,8).readUInt16BE(0);
-    let correlation_bytes = data.subarray(8,12).readUInt16BE(0);
+    let correlation_bytes = data.subarray(8,12).readUInt32BE(0);
 
     console.log(api_version_bytes);
     console.log(correlation_bytes);
